fix cg and other error

+++ chulkov.leontii/I4/fillingthematrixinspiral.cpp
@@ -1,27 +1,28 @@
 #include <cstddef>
+#include "fillingthematrixinspiral.h"
 
 int* fillingTheMatrixInSpiral(int* matrix, size_t c, size_t r, size_t indrow, size_t indcol)
 {
   indrow--;
   indcol--;
   matrix[c * indrow + indcol] = 1;
-  int val = 1;
-  int top = 0, bottom = r - 1, left = 0, right = c - 1;
+  size_t val = 1;
+  size_t top = 0, bottom = r - 1, left = 0, right = c - 1;
   while (top <= bottom && left <= right)
   {
-    for (int i = top; i <= bottom; i++)
+    for (size_t i = top; i <= bottom; i++)
     {
       matrix[c * i + left] = val++;
     }
     left++;
-    for (int i = left; i <= right; i++)
+    for (size_t i = left; i <= right; i++)
     {
       matrix[c * bottom + i] = val++;
     }
     bottom--;
     if (left <= right)
     {
-      for (int i = bottom; i >= top; i--)
+      for (size_t i = bottom; i >= top; i--)
       {
         matrix[c * i + right] = val++;
       }
@@ -29,7 +30,7 @@ int* fillingTheMatrixInSpiral(int* matrix, size_t c, size_t r, size_t indrow, si
     }
     if (top <= bottom)
     {
-      for (int i = right; i >= left; i--)
+      for (size_t i = right; i >= left; i--)
       {
         matrix[c * top + i] = val++;
       }
